decimals are available in floats, not longs...
git-svn-id: svn://oss.beweb.dbc.dk/sync_search/opensearch/trunk@2346 51cac136-f83d-432d-bb6f-125e83a60a2d

+++ src/dbc/opensearch/components/datadock/tests/datadocktest/DataDockShmTest.java
@@ -50,7 +50,7 @@ public class DataDockShmTest {
             e.printStackTrace();
         }
         
-        Long estimate=null;
+        Float estimate=null;
         try{
             estimate = dd.call(); 
         }catch(Exception e){
